Refactored Code, Added Dealer class
Made the classes work with each other a lot nicer and put methods inside of others so they wouldn't need to manually be called anymore.

I also added Dealer class for dealer behaviors, such as hiding the first card.

Also added Characters.py to separate the classes for the Player and Dealer with the Deck